+++ CHANGELOG.md
@@ -5,6 +5,16 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.1.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
+## [1.8.4] - 2026-08-01
+
+### Fixed
+
+- Fixed issues with variant-specific image overrides.
+
+### Changed
+
+- Improvements to resource pack override IDs.
+
 ## [1.8.3] - 2026-06-15
 
 ### Fixed